refactor(api): move controller and test to Team domain
