+++ api/tests/unit/application/users/user-controller_test.js
@@ -1,5 +1,4 @@
 import { userController } from '../../../../lib/application/users/user-controller.js';
-import { usecases } from '../../../../lib/domain/usecases/index.js';
 import { usecases as devcompUsecases } from '../../../../src/devcomp/domain/usecases/index.js';
 import { expect, hFake, sinon } from '../../../test-helper.js';
 
@@ -44,39 +43,6 @@ describe('Unit | Controller | user-controller', function () {
       });
       expect(trainingSerializer.serialize).to.have.been.calledWithExactly(userRecommendedTrainings, meta);
       expect(response).to.equal(expectedResult);
-    });
-  });
-
-  describe('#findCertificationCenterMembershipsByUser', function () {
-    it("should return user's certification centers", async function () {
-      // given
-      const certificationCenterMemberships = Symbol("a list of user's certification center memberships");
-      const certificationCenterMembershipsSerialized = Symbol(
-        "a list of user's certification center memberships serialized",
-      );
-
-      const certificationCenterMembershipSerializer = { serializeForAdmin: sinon.stub() };
-      certificationCenterMembershipSerializer.serializeForAdmin
-        .withArgs(certificationCenterMemberships)
-        .returns(certificationCenterMembershipsSerialized);
-
-      sinon
-        .stub(usecases, 'findCertificationCenterMembershipsByUser')
-        .withArgs({ userId: 12345 })
-        .resolves(certificationCenterMemberships);
-
-      // when
-      const request = {
-        params: {
-          id: 12345,
-        },
-      };
-      const result = await userController.findCertificationCenterMembershipsByUser(request, hFake, {
-        certificationCenterMembershipSerializer,
-      });
-
-      // then
-      expect(result.source).to.equal(certificationCenterMembershipsSerialized);
     });
   });
 });
